# schemas/openapi.yaml
openapi: "3.1.0"
info:
  title: POLARIS API
  version: "0.1.0"
  description: |
    Decision-support API. Not an official warning service.
    IMPLEMENTED: GET /health and flood slice routes over SIMULATED and
    HISTORICAL_REPLAY fixtures. LIVE adapters, CHI, HCI engine, and
    OFFICIAL/CAP alerting are NOT IMPLEMENTED.
paths:
  /health:
    get:
      summary: Liveness/readiness
      operationId: getHealth
      responses:
        "200":
          description: Service is up
          content:
            application/json:
              schema:
                type: object
                required: [status, service, evidence]
                properties:
                  status: { type: string, example: ok }
                  service: { type: string, example: polaris }
                  evidence: { type: string, example: IMPLEMENTED }
                  maturity: { type: string }
                  utc: { type: string, format: date-time }
  /v1/observations:
    get:
      summary: List observations used by the flood slice
      operationId: listObservations
      parameters:
        - in: query
          name: seed
          schema: { type: integer, default: 42 }
        - in: query
          name: fixture_id
          schema: { type: string, default: flood-bogota-demo }
      responses:
        "200":
          description: Observation list (SIMULATED or HISTORICAL_REPLAY)
  /v1/observations/{observation_id}:
    get:
      summary: Get one observation
      operationId: getObservation
      parameters:
        - in: path
          name: observation_id
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200": { description: Observation }
        "404": { description: Not found }
  /v1/assessments:
    get:
      summary: Run the flood slice (observation→GCI→PHI→risk→DRAFT alert)
      operationId: listAssessments
      responses:
        "200":
          description: Assessments; alerts are DRAFT only
  /v1/assessments/{spatial_unit_id}:
    get:
      summary: Assessment for one spatial unit
      operationId: getAssessment
      parameters:
        - in: path
          name: spatial_unit_id
          required: true
          schema: { type: string }
      responses:
        "200": { description: Assessment }
        "404": { description: Not found }
  /v1/assessments/run:
    post:
      summary: Re-run the flood slice with a seed
      operationId: runAssessments
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                seed: { type: integer, default: 42 }
                fixture_id: { type: string, default: flood-bogota-demo }
                data_class: { type: string, enum: [SIMULATED, HISTORICAL_REPLAY] }
      responses:
        "200": { description: Slice result }
        "400": { description: LIVE or mismatched data_class refused }
  /v1/ingest/iot:
    post:
      summary: Run SIMULATED IoT scenario through gateway ingest and flood slice
      operationId: ingestIot
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                seed: { type: integer, default: 42 }
                scenario_id: { type: string, default: iot-bogota-demo }
      responses:
        "200": { description: IoT ingest + flood_slice (SIMULATED only) }
        "404": { description: Scenario not found }
  /v1/alerts:
    get:
      summary: DRAFT decision-support alerts (never OFFICIAL)
      operationId: listAlerts
      responses:
        "200": { description: DRAFT alerts }
  /v1/map/geojson:
    get:
      summary: GeoJSON FeatureCollection for Horizon
      operationId: mapGeojson
      responses:
        "200": { description: FeatureCollection labelled SIMULATED or HISTORICAL_REPLAY }
  /v1/backtests/{scenario_id}:
    get:
      summary: HISTORICAL_REPLAY hit/miss report
      operationId: getBacktest
      parameters:
        - in: path
          name: scenario_id
          required: true
          schema: { type: string }
        - in: query
          name: seed
          schema: { type: integer, default: 42 }
      responses:
        "200": { description: EXPERIMENTAL backtest JSON }
        "404": { description: Scenario not found }